refactor(mcp): extract rollback tool module
