+++ openspec/changes/refactor-mcp-rollback-tool-module/README.md
@@ -1,0 +1,3 @@
+# refactor-mcp-rollback-tool-module
+
+Refactor: extract MCP rollback tool implementation into a dedicated module
