+++ src/mcp/tools.rs
@@ -23,6 +23,7 @@ use super::{AgentpackMcp, confirm};
 mod evolve_propose;
 mod evolve_restore;
 mod explain;
+mod rollback;
 
 pub(super) const TOOLS_INSTRUCTIONS: &str = "Agentpack MCP server (stdio). Tools: plan, diff, preview, status, doctor, deploy, deploy_apply, rollback, evolve_propose, evolve_restore, explain.";
 
@@ -486,43 +487,7 @@ async fn call_explain_in_process(args: ExplainArgs) -> anyhow::Result<(String, s
 async fn call_rollback_in_process(
     args: RollbackArgs,
 ) -> anyhow::Result<(String, serde_json::Value)> {
-    tokio::task::spawn_blocking(move || {
-        let home = crate::paths::AgentpackHome::resolve().context("resolve agentpack home")?;
-
-        let RollbackArgs {
-            repo: _repo_override,
-            to: snapshot_id,
-            yes,
-        } = args;
-        let result = crate::handlers::rollback::rollback(&home, &snapshot_id, true, yes);
-
-        let (text, envelope) = match result {
-            Ok(event) => {
-                let data = crate::app::rollback_json::rollback_json_data(&snapshot_id, &event.id);
-                let envelope = crate::output::JsonEnvelope::ok("rollback", data);
-                let text = serde_json::to_string_pretty(&envelope)?;
-                let envelope = serde_json::to_value(&envelope)?;
-                (text, envelope)
-            }
-            Err(err) => {
-                let user_err = err.chain().find_map(|e| e.downcast_ref::<UserError>());
-                let code = user_err
-                    .map(|e| e.code.clone())
-                    .unwrap_or_else(|| "E_UNEXPECTED".to_string());
-                let message = user_err
-                    .map(|e| e.message.clone())
-                    .unwrap_or_else(|| err.to_string());
-                let details = user_err.and_then(|e| e.details.clone());
-                let envelope = envelope_error("rollback", &code, &message, details);
-                let text = serde_json::to_string_pretty(&envelope)?;
-                (text, envelope)
-            }
-        };
-
-        Ok((text, envelope))
-    })
-    .await
-    .context("mcp rollback handler task join")?
+    rollback::call_rollback_in_process(args).await
 }
 
 async fn call_evolve_restore_in_process(
